user_service/pyproject.toml
[tool.poetry]
name = "user_service"
version = "0.1.0"
description = ""
authors = ["Bohdan Salamakha <allen.avanheim@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
pip = "^24.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
psycopg2-binary = "^2.9.9"
setuptools = "^69.5.1"
gunicorn = "^22.0.0"
pika = "^1.3.2"
pika-stubs = "^0.1.3"
gevent = "^24.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
